requirement.txt
pandas
numpy
scikit-learn
azureml-core
